analyser/requirements.txt
aioredis==2.0.1
anyio==3.6.1
async-timeout==4.0.2
attrs==22.1.0
click==8.1.3
fastapi==0.85.0
h11==0.13.0
httptools==0.5.0
idna==3.4
iniconfig==1.1.1
motor==3.0.0
nats-py==2.1.7
packaging==21.3
pluggy==1.0.0
py==1.11.0
pydantic==1.10.2
pymongo==4.2.0
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==0.21.0
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.3.0
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
websockets==10.3
